requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
torch
fastapi
uvicorn
jinja2
qdrant-client
sentence-transformers
numpy
python-multipart
python-dotenv
google-generativeai
markdown
